Refactor tests Fix wrong exception in validate method
